fix request proto conversions and rename package
