+++ plugins/transport-grpc/build.gradle
@@ -12,7 +12,7 @@ apply plugin: 'opensearch.internal-cluster-test'
 
 opensearchplugin {
   description = 'gRPC based transport implementation'
-  classname = 'org.opensearch.transport.grpc.GrpcPlugin'
+  classname = 'org.opensearch.plugin.transport.grpc.GrpcPlugin'
 }
 
 dependencies {
