+++ plugins/transport-grpc/src/main/java/org/opensearch/plugin/transport/grpc/proto/package-info.java
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * This package contains Protocol Buffer conversion methods for OpenSearch objects.
+ * These methods aim to centralize all Protocol Buffer conversion logic in the transport-grpc module.
+ */
+package org.opensearch.plugin.transport.grpc.proto;
