--- plugins/transport-grpc/src/main/java/org/opensearch/transport/grpc/package-info.java
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * gRPC transport implementation for OpenSearch.
- * Provides network communication using the gRPC protocol.
- */
-package org.opensearch.transport.grpc;
