--- plugins/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/package-info.java
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * This package contains Protocol Buffer conversion methods for OpenSearch objects.
- * These methods aim to centralize all Protocol Buffer conversion logic in the transport-grpc module.
- */
-package org.opensearch.transport.grpc.proto;
