--- plugins/transport-grpc/src/test/java/org/opensearch/transport/grpc/GrpcPluginTests.java
@@ -1,97 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.transport.grpc;
-
-import org.opensearch.common.network.NetworkService;
-import org.opensearch.common.settings.ClusterSettings;
-import org.opensearch.common.settings.Setting;
-import org.opensearch.common.settings.Settings;
-import org.opensearch.core.indices.breaker.CircuitBreakerService;
-import org.opensearch.plugins.NetworkPlugin;
-import org.opensearch.telemetry.tracing.Tracer;
-import org.opensearch.test.OpenSearchTestCase;
-import org.opensearch.threadpool.ThreadPool;
-import org.junit.Before;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-
-import java.util.List;
-import java.util.Map;
-import java.util.function.Supplier;
-
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.GRPC_TRANSPORT_SETTING_KEY;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_BIND_HOST;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_HOST;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_PORT;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_PUBLISH_HOST;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_PUBLISH_PORT;
-import static org.opensearch.transport.grpc.Netty4GrpcServerTransport.SETTING_GRPC_WORKER_COUNT;
-
-public class GrpcPluginTests extends OpenSearchTestCase {
-
-    private GrpcPlugin plugin;
-
-    @Mock
-    private ThreadPool threadPool;
-
-    @Mock
-    private CircuitBreakerService circuitBreakerService;
-
-    @Mock
-    private NetworkService networkService;
-
-    @Mock
-    private ClusterSettings clusterSettings;
-
-    @Mock
-    private Tracer tracer;
-
-    @Before
-    public void setup() {
-        MockitoAnnotations.openMocks(this);
-        plugin = new GrpcPlugin();
-    }
-
-    public void testGetSettings() {
-        List<Setting<?>> settings = plugin.getSettings();
-
-        // Verify that all expected settings are returned
-        assertTrue("SETTING_GRPC_PORT should be included", settings.contains(SETTING_GRPC_PORT));
-        assertTrue("SETTING_GRPC_HOST should be included", settings.contains(SETTING_GRPC_HOST));
-        assertTrue("SETTING_GRPC_PUBLISH_HOST should be included", settings.contains(SETTING_GRPC_PUBLISH_HOST));
-        assertTrue("SETTING_GRPC_BIND_HOST should be included", settings.contains(SETTING_GRPC_BIND_HOST));
-        assertTrue("SETTING_GRPC_WORKER_COUNT should be included", settings.contains(SETTING_GRPC_WORKER_COUNT));
-        assertTrue("SETTING_GRPC_PUBLISH_PORT should be included", settings.contains(SETTING_GRPC_PUBLISH_PORT));
-
-        // Verify the number of settings
-        assertEquals("Should return 6 settings", 6, settings.size());
-    }
-
-    public void testGetAuxTransports() {
-        Settings settings = Settings.builder()
-            .put(SETTING_GRPC_PORT.getKey(), "9200-9300")
-            .build();
-
-        Map<String, Supplier<NetworkPlugin.AuxTransport>> transports = plugin.getAuxTransports(
-            settings,
-            threadPool,
-            circuitBreakerService,
-            networkService,
-            clusterSettings,
-            tracer
-        );
-
-        // Verify that the transport map contains the expected key
-        assertTrue("Should contain GRPC_TRANSPORT_SETTING_KEY", transports.containsKey(GRPC_TRANSPORT_SETTING_KEY));
-
-        // Verify that the supplier returns a Netty4GrpcServerTransport instance
-        NetworkPlugin.AuxTransport transport = transports.get(GRPC_TRANSPORT_SETTING_KEY).get();
-        assertTrue("Should return a Netty4GrpcServerTransport instance", transport instanceof Netty4GrpcServerTransport);
-    }
-}
